+++ src/chopstick/chop_utils.h
@@ -21,17 +21,17 @@ vector<vector<int>> all_permutations(vector<int> s_A);
 
 vector<vector<int>> divides_strategies(int resources, int fields_number);
 
-long int k_W(vector<int> s_A, vector<int> s_B);
+long long k_W(vector<int> s_A, vector<int> s_B);
 
-long int k_L(vector<int> s_A, vector<int> s_B);
+long long k_L(vector<int> s_A, vector<int> s_B);
 
-long int factorial(int n);
+long long factorial(int n);
 
 int matrix_sum(vector<vector<int>> matrix);
 
 vector<int> find_area_vector(vector<vector<int>> clash_mat, int area);
 
-vector<vector<long int>> empty_matrix(int size);
+vector<vector<long long>> empty_matrix(int size);
 
 vector<vector<int>> empty_clash_matrix(int size);
 
@@ -49,9 +49,9 @@ vector<int> trim_vector(vector<int> v, int number_of_elements);
 
 vector<vector<int>> find_knots(vector<int> L, vector<int> T);
 
-long int newton_symbol(int n, int k);
+long long newton_symbol(int n, int k);
 
-long int single_type_rectangle(int cols_num, int rows_num, int rooks_num);
+long long single_type_rectangle(int cols_num, int rows_num, int rooks_num);
 
 int min_number_of_rooks(int i, int j, int n);
